Convert Elixir project to umbrella structure
Move the cli/ directory to apps/cli/ to enable future multi-app development.
This prepares shimmer for having multiple CLI tools and libraries that can
share dependencies and be developed together.

Changes:
- Add umbrella root mix.exs with shared release configuration
- Move cli/ to apps/cli/ with updated paths for umbrella child app
- Create root config/config.exs for shared configuration
- Move .formatter.exs and .credo.exs to root level
- Update CI workflow to use root-level paths (_build, deps, mix.lock)
- Update mise tasks to run from umbrella root

Fixes #409
