+++ .formatter.exs
@@ -1,0 +1,10 @@
+# Used by "mix format"
+[
+  inputs: [
+    "{mix,.formatter}.exs",
+    "config/*.exs",
+    "apps/*/{mix,.formatter}.exs",
+    "apps/*/config/*.exs",
+    "apps/*/{lib,test}/**/*.{ex,exs}"
+  ]
+]
