+++ apps/cli/.gitignore
