+++ apps/cli/priv/prompts/agents/brownie.txt
