+++ apps/cli/priv/prompts/agents/johnson.txt
