+++ apps/cli/priv/prompts/agents/junior.txt
