+++ apps/cli/priv/prompts/agents/k7r2.txt
