+++ apps/cli/priv/prompts/agents/quick.txt
