+++ apps/cli/priv/prompts/agents/rho.txt
