+++ apps/cli/test/cli_test.exs
