+++ apps/cli/test_streaming.exs
