+++ config/config.exs
@@ -1,0 +1,7 @@
+# This file is responsible for configuring your umbrella
+# and **all applications** and their dependencies with the
+# temporary exception of runtime configuration.
+#
+# Configuration specific to individual apps can be found in
+# each application's `config/` directory (if they have one).
+import Config
